first_ai_short_form-main/step1_Shot_Boundary_Detection/requirements.txt
# requirements.txt
torch
torchvision
opencv-python
pyyaml
